Added test job to the enterprise repo
