+++ job/testjob/testjob.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-package job
+package testjob
 
 import (
 	"time"
@@ -9,11 +9,6 @@ import (
 	l4g "github.com/alecthomas/log4go"
 	"github.com/mattermost/platform/store"
 )
-
-type Job interface {
-	Run()
-	Stop()
-}
 
 type TestJob struct {
 	store store.Store
